+++ NEWS.md
@@ -1,8 +1,7 @@
 ## 2018-09-19 v1.0.0
  * Add schema description to create validation_rules table
  * Add endpoints /tenant/rules with GET, POST and PUT methods to manage rules for tenant
- * Add endpoint /password/validators to get active rules by type
- * Add endpoint /password/validate for password validation
+ * Add endpoint /validate for password validation
  * Implement Validation Engine and Validation Registry services
  * Set up endpoint permissions 
 ## 2018-09-04 v0.0.1